# shared/api/bundles/order.openapi.v1.bundle.yaml
openapi: 3.0.3
info:
  title: OrderServiceApi
  version: 1.0.0
  description: Api для работы с заказами
tags:
  - name: Order
    description: Операции с заказами. Создание, получение, оплата, отмена
paths:
  /api/v1/orders:
    post:
      summary: Create new order
      operationId: NewOrder
      tags:
        - Order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/create_order_request'
      responses:
        '200':
          description: Заказ успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/create_order_response'
        '404':
          description: Детали в заказе не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/not_found_error'
        '500':
          description: Внутрення ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_server_error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
  /api/v1/orders/{order_uuid}/pay:
    parameters:
      - $ref: '#/components/parameters/order_uuid'
    post:
      summary: Оплата заказа
      operationId: PayOrder
      tags:
        - Order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/pay_order_request'
      responses:
        '200':
          description: Order information successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pay_order_response'
        '404':
          description: Order for uuid not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/not_found_error'
        '409':
          description: Order for uuid not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/conflict_error'
        '500':
          description: Order for uuid not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_server_error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
  /api/v1/orders/{order_uuid}:
    parameters:
      - $ref: '#/components/parameters/order_uuid'
    get:
      summary: Get Order data by uuid
      operationId: GetOrderByUUID
      tags:
        - Order
      responses:
        '200':
          description: Order information successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/order_dto'
        '404':
          description: Order for uuid not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/not_found_error'
        '500':
          description: Order for uuid not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_server_error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
  /api/v1/orders/{order_uuid}/cancel:
    parameters:
      - $ref: '#/components/parameters/order_uuid'
    post:
      summary: Cancel order
      operationId: CancelOrder
      tags:
        - Order
      responses:
        '204':
          description: Заказ успешно отменен
        '404':
          description: Детали в заказе не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/not_found_error'
        '409':
          description: Заказ уже оплачен и не может быть отменен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/conflict_error'
        '500':
          description: Внутрення ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_server_error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
components:
  schemas:
    create_order_request:
      type: object
      required:
        - user_uuid
        - part_uuids
      properties:
        user_uuid:
          type: string
          format: uuid
          description: Уникальный идентификатор пользователя
        part_uuids:
          type: array
          description: Массив уникальный идентификаторов запчастейЭ
          items:
            type: string
            format: uuid
    generic_error:
      type: object
      properties:
        code:
          type: integer
          description: HTTP Код ошибки
        message:
          type: string
          description: Описание ошибки
    create_order_response:
      type: object
      required:
        - order_uuid
        - total_price
      properties:
        uuid:
          type: string
          format: uuid
          description: Уникальный идентификатор заказа
        total_price:
          type: number
          format: float64
          description: Стоимость заказа
    not_found_error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP Код ошибки
          example: 404
        message:
          type: string
          description: Описание ошибки
          example: Order by uuid not found
    internal_server_error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP Код ошибки
          example: 500
        message:
          type: string
          description: Описание ошибки
          example: Internal server error
    payment_method:
      type: string
      description: Тип оплаты
      enum:
        - PAYMENT_METHOD_UNKNOWN
        - PAYMENT_METHOD_CARD
        - PAYMENT_METHOD_SBP
        - PAYMENT_METHOD_CREDIT_CARD
        - PAYMENT_METHOD_INVESTOR_MONEY
    pay_order_request:
      type: object
      required:
        - payment_method
      properties:
        payment_method:
          $ref: '#/components/schemas/payment_method'
    pay_order_response:
      type: object
      required:
        - transaction_uuid
      properties:
        transaction_uuid:
          type: string
          format: uuid
          description: Уникальный идентификатор транзакции
    conflict_error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP Код ошибки
          example: 409
        message:
          type: string
          description: Описание ошибки
          example: Conflict
    order_status:
      type: string
      description: Статус заказа
      enum:
        - PENDING_PAYMENT
        - PAID
        - CANCELLED
    order_dto:
      type: object
      required:
        - order_uuid
        - user_uuid
        - part_uuids
        - total_price
        - status
      properties:
        order_uuid:
          type: string
          format: uuid
          description: Уникальный идентификатор заказа
          minLength: 1
          maxLength: 100
        user_uuid:
          type: string
          format: uuid
          description: Уникальный идентификатор пользователя создавшего заказ
          minLength: 1
          maxLength: 100
        part_uuids:
          type: array
          description: Список uuid запчастей в заказе
          minItems: 1
          items:
            type: string
            format: uuid
        total_price:
          type: number
          description: Итоговая стоимость заказа
          format: double
          example: 12.34
        transaction_uuid:
          type: string
          format: uuid
          description: UUID транзакции (если оплачен)
        payment_method:
          type: string
          description: Способ оплаты (если оплачен)
        status:
          $ref: '#/components/schemas/order_status'
  parameters:
    order_uuid:
      name: order_uuid
      in: path
      required: true
      description: Универсальный идентификатор заказа
      schema:
        type: string
        format: uuid
        minLength: 0
        maxLength: 100
x-ogen:
  target: ./shared/pkg/openapi/order/v1
  package: order_v1
  clean: true